+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-# v1.0.0
+# v1.0.1
     -initial release
+
+# v1.0.2
+    -Get Message Count
+    -Delete all messages